requirements.txt
h5py
einops
accelerate
ema_pytorch
tensorboardX
termcolor
protobuf==3.20.*
einops-exts
rotary-embedding-torch
shapely
OApackage==2.7.6
pyrallis==0.3.1

Keras==2.13.1
numpy==1.22.4
scipy==1.7.3
tensorflow==2.13.1
tensorboard

json5==0.9.5
matplotlib==3.5.1
datasets==3.1.0